fix: move expense participant projection to repo layer
