+++ backend/tevent/src/main/java/com/tbank/tevent/repo/ExpenseParticipantView.java
@@ -1,4 +1,4 @@
-package com.tbank.tevent.expenses;
+package com.tbank.tevent.repo;
 
 import java.util.UUID;
